feat: allow query cancellation
Add a second function `ExecuteWithContext` which allows passing a
`net.Context` with a query. This context can be canceled, which in turn
cancels calling all of the remaining resolver functions. Furthermore,
resolver functions can also handle cancellation of the context.

This has the added benefit of guaranteeing all resolvers will be
canceled if the main Execute function returns early for any reason.

Signed-off-by: Christian Stewart <christian@paral.in>